Move CacheMissError into separate file
